mod6_labs/requirements.txt
anyio==4.11.0
certifi==2025.11.12
flet==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
python-dotenv==1.2.1
repath==0.9.0
six==1.17.0
sniffio==1.3.1
